requirements-v3.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
cachetools==5.3.0
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==3.1.0
click==8.0.3
cloud-init==22.4.2
cloud-tpu-client==0.10
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cramjam==2.6.2
crcmod==1.7
cryptography==3.4.8
Cython==0.29.28
dbus-python==1.2.18
decorator==5.1.1
distlib==0.3.6
distro==1.7.0
distro-info===1.1build1
fastparquet==2023.2.0
filelock==3.10.7
flatbuffers==23.3.3
frozenlist==1.3.3
fsspec==2023.4.0
future==0.18.3
gast==0.4.0
gcsfs==2023.4.0
google-api-core==1.34.0
google-api-python-client==1.8.0
google-auth==2.17.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.3.2
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
grpcio==1.53.0
h5py==3.8.0
httplib2==0.22.0
huggingface-hub==0.13.4
hyperlink==21.0.0
idna==3.4
importlib-metadata==4.6.4
incremental==21.3.0
intel-openmp==2023.1.0
jax==0.4.8
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.2.0
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keras==2.12.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
mkl==2023.1.0
mkl-include==2023.1.0
ml-dtypes==0.0.4
mock==5.0.1
more-itertools==8.10.0
multidict==6.0.4
netifaces==0.11.0
nltk==3.8.1
numpy==1.23.5
oauth2client==4.1.3
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==2.0.0
pexpect==4.8.0
platformdirs==3.2.0
protobuf==3.20.3
ptyprocess==0.7.0
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==3.0.9
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-dotenv==1.0.0
python-magic==0.4.24
pytz==2022.1
PyYAML==5.4.1
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
rouge-score==0.1.2
rsa==4.9
scipy==1.10.1
SecretStorage==3.3.1
sentencepiece==0.1.98
service-identity==18.1.0
six==1.16.0
sos==4.4
ssh-import-id==5.11
systemd-python==234
tbb==2021.9.0
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.12.0
tensorflow-hub==0.13.0
tensorflow-io-gcs-filesystem==0.32.0
tensorflow-text==2.12.0
termcolor==2.2.0
tokenizers==0.13.3
tqdm==4.65.0
transformers==4.28.1
Twisted==22.1.0
typing_extensions==4.5.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.26.15
virtualenv==20.21.0
wadllib==1.3.6
Werkzeug==2.2.3
wrapt==1.14.1
yarl==1.8.2
zipp==1.0.0
zope.interface==5.4.0
